requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
manga-ocr==0.1.11
pillow==10.2.0